test: get one event via path and query param is identical
